Split benchmark in multiple file to allow benchmark without optional features
