+++ Cargo.toml
@@ -27,7 +27,20 @@ xz2 = { version = "0.1", optional = true }
 tempfile = "3"
 criterion = "0.3"
 
+[[bench]]
+name = "detect_format"
+harness = false
 
 [[bench]]
-name = "benchmarks"
+name = "gzip"
 harness = false
+
+[[bench]]
+name = "bzip2"
+harness = false
+required-features = ["bz2"]
+
+[[bench]]
+name = "lzma"
+harness = false
+required-features = ["xz2"]
